feat(web): audit goes standalone — /[org]/[repo]/audit, chapter sidebar, file preview
- AuditWorkspace extracted from AuditPanel (thin wrapper keeps the
  playground integration) so the same experience runs in both homes
- standalone /audit page inside the standard [org]/[repo] layout shell:
  repo prefilled from URL, app Navbar/Sidebar chrome for free
- Sidebar gains an Audit entry
- chapter jump sidebar above the fly-cam: counts per chapter, click
  jumps the walkthrough cursor to the chapter's first finding and
  scrolls the transcript to it
- 'lihat file ↗' button on every finding with a location: overlay
  fetches /api/file (raw.githubusercontent, no clone), line numbers,
  the finding's line highlighted red and scrolled into view, esc closes
